:recycle: [Refactor] 컨트롤러, 서비스, 시큐리티 폴더, 파일 정리
