+++ src/main/java/com/gamegoo/security/CustomUserDetails.java
@@ -1,6 +1,7 @@
 package com.gamegoo.security;
 
 import com.gamegoo.domain.Member;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -8,13 +9,9 @@ import java.util.ArrayList;
 import java.util.Collection;
 
 // UserDetails 오버라이딩해서 Spring Security의 Login Filter에 사용하는 클래스
+@RequiredArgsConstructor
 public class CustomUserDetails implements UserDetails {
     private final Member member;
-
-    public CustomUserDetails(Member member) {
-
-        this.member = member;
-    }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
