+++ src/main/java/com/gamegoo/service/member/CustomUserDetailService.java
@@ -1,10 +1,11 @@
-package com.gamegoo.security;
+package com.gamegoo.service.member;
 
-import com.gamegoo.apiPayload.code.status.ErrorStatus;
 import com.gamegoo.apiPayload.exception.CustomUserException;
-import com.gamegoo.apiPayload.exception.handler.MemberHandler;
 import com.gamegoo.domain.Member;
 import com.gamegoo.repository.member.MemberRepository;
+import com.gamegoo.security.CustomUserDetails;
+import io.jsonwebtoken.JwtException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -12,13 +13,9 @@ import org.springframework.stereotype.Service;
 
 // 로그인 DB 로직
 @Service
+@RequiredArgsConstructor
 public class CustomUserDetailService implements UserDetailsService {
     private final MemberRepository memberRepository;
-
-    public CustomUserDetailService(MemberRepository memberRepository) {
-
-        this.memberRepository = memberRepository;
-    }
 
     @Override
     public UserDetails loadUserByUsername(String email) {
@@ -29,13 +26,13 @@ public class CustomUserDetailService implements UserDetailsService {
                     }
                     return new CustomUserDetails(member);
                 })
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new JwtException("해당 사용자를 찾을 수 없습니다."));
     }
 
     public UserDetails loadUserById(Long id) throws UsernameNotFoundException {
         Member member = memberRepository.findById(id)
                 .filter(m -> !m.getBlind())
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new JwtException("해당 사용자를 찾을 수 없습니다."));
         return new CustomUserDetails(member);
     }
