--- src/main/java/com/gamegoo/service/member/DeleteService.java
@@ -1,33 +1,0 @@
-package com.gamegoo.service.member;
-
-import com.gamegoo.apiPayload.code.status.ErrorStatus;
-import com.gamegoo.apiPayload.exception.handler.MemberHandler;
-import com.gamegoo.domain.Member;
-import com.gamegoo.repository.member.MemberRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class DeleteService {
-    private final MemberRepository memberRepository;
-
-    @Autowired
-
-    public DeleteService(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    public void deleteMember(Long userId) {
-        Optional<Member> optionalMember = memberRepository.findById(userId);
-        if (optionalMember.isPresent()) {
-            Member member = optionalMember.get();
-            member.setBlind(true);
-            memberRepository.save(member);
-        } else {
-            throw new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND);
-        }
-
-    }
-}
