+++ src/main/java/com/gamegoo/service/member/ProfileService.java
@@ -1,17 +1,17 @@
 package com.gamegoo.service.member;
 
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.gamegoo.apiPayload.ApiResponse;
 import com.gamegoo.apiPayload.code.status.ErrorStatus;
+import com.gamegoo.apiPayload.exception.handler.MemberHandler;
 import com.gamegoo.domain.Member;
 import com.gamegoo.domain.gamestyle.GameStyle;
 import com.gamegoo.domain.gamestyle.MemberGameStyle;
 import com.gamegoo.repository.member.GameStyleRepository;
 import com.gamegoo.repository.member.MemberGameStyleRepository;
 import com.gamegoo.repository.member.MemberRepository;
-import com.gamegoo.security.SecurityUtil;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.gamegoo.util.SecurityUtil;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,26 +23,19 @@ import java.util.Optional;
 import java.util.Set;
 
 @Service
-public class MemberGameStyleService {
-
+@RequiredArgsConstructor
+public class ProfileService {
     private final MemberRepository memberRepository;
     private final GameStyleRepository gameStyleRepository;
     private final MemberGameStyleRepository memberGameStyleRepository;
     private final HttpServletResponse response;
 
-    @Autowired
-    public MemberGameStyleService(MemberRepository memberRepository, GameStyleRepository gameStyleRepository, MemberGameStyleRepository memberGameStyleRepository, HttpServletResponse response) {
-        this.memberRepository = memberRepository;
-        this.gameStyleRepository = gameStyleRepository;
-        this.memberGameStyleRepository = memberGameStyleRepository;
-        this.response = response;
-    }
 
     @Transactional
     public void addMemberGameStyles(List<String> gameStyles) throws IOException {
         Long userId = SecurityUtil.getCurrentUserId();
         ErrorStatus errorStatus = null;
-        
+
         Optional<Member> member = memberRepository.findById(userId);
 
         if (member.isEmpty()) {
@@ -91,6 +84,79 @@ public class MemberGameStyleService {
                 }
 
             }
+        }
+    }
+
+    public void deleteMember(Long userId) {
+        Optional<Member> optionalMember = memberRepository.findById(userId);
+        if (optionalMember.isPresent()) {
+            Member member = optionalMember.get();
+            member.setBlind(true);
+            memberRepository.save(member);
+        } else {
+            throw new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND);
+        }
+
+    }
+
+    public void modifyPosition(int mainP, int subP) throws IOException {
+        Long userId = SecurityUtil.getCurrentUserId();
+        ErrorStatus errorStatus = null;
+
+        // 만약 mainP, subP 가 1~5의 값이 아닐 경우
+        if (mainP <= 0 || subP <= 0 || mainP > 5 || subP > 5) {
+            errorStatus = ErrorStatus.POSITION_NOT_FOUND;
+        }
+
+        Optional<Member> optionalMember = memberRepository.findById(userId);
+
+        if (optionalMember.isPresent()) {
+            Member member = optionalMember.get();
+            member.setMainPosition(mainP);
+            member.setSubPosition(subP);
+            memberRepository.save(member);
+        } else {
+            errorStatus = ErrorStatus.MEMBER_NOT_FOUND;
+        }
+
+        if (errorStatus != null) {
+            ApiResponse<Object> apiResponse = ApiResponse.onFailure(errorStatus.getCode(), errorStatus.getMessage(), null);
+
+            response.setStatus(errorStatus.getHttpStatus().value());
+            response.setContentType("application/json");
+            response.setCharacterEncoding("UTF-8");
+            new ObjectMapper().writeValue(response.getWriter(), apiResponse);
+
+        }
+    }
+
+    public void modifyProfileImage(String profileImage) throws IOException {
+        Long userId = SecurityUtil.getCurrentUserId();
+        ErrorStatus errorStatus = null;
+
+        if (profileImage.length() > 30) {
+            errorStatus = ErrorStatus.PROFILE_IMAGE_BAD_REQUEST;
+        } else {
+            Optional<Member> optionalMember = memberRepository.findById(userId);
+
+            if (optionalMember.isPresent()) {
+                Member member = optionalMember.get();
+                member.setProfileImage(profileImage);
+                memberRepository.save(member);
+            } else {
+                errorStatus = ErrorStatus.MEMBER_NOT_FOUND;
+            }
+        }
+
+
+        if (errorStatus != null) {
+            ApiResponse<Object> apiResponse = ApiResponse.onFailure(errorStatus.getCode(), errorStatus.getMessage(), null);
+
+            response.setStatus(errorStatus.getHttpStatus().value());
+            response.setContentType("application/json");
+            response.setCharacterEncoding("UTF-8");
+            new ObjectMapper().writeValue(response.getWriter(), apiResponse);
+
         }
     }
 }
